Remove now-bogus (uint16) casts when storing AST node line numbers (patch revised based on ed@siliconforks.com's patch in 500128, r=me).

+++ js/src/jsscan.h
@@ -287,7 +287,7 @@ struct JSTokenStream {
 };
 
 #define CURRENT_TOKEN(ts)       ((ts)->tokens[(ts)->cursor])
-#define ON_CURRENT_LINE(ts,pos) ((uint16)(ts)->lineno == (pos).end.lineno)
+#define ON_CURRENT_LINE(ts,pos) ((ts)->lineno == (pos).end.lineno)
 
 /* JSTokenStream flags */
 #define TSF_ERROR       0x01            /* fatal error while compiling */
